+++ logtape/record.ts
@@ -23,6 +23,21 @@ export interface LogRecord {
   readonly message: readonly unknown[];
 
   /**
+   * The raw log message.  This is the original message template without any
+   * further processing.  It can be either:
+   *
+   * - A string without any substitutions if the log record was created with
+   *   a method call syntax, e.g., "Hello, {name}!" for
+   *   `logger.info("Hello, {name}!", { name })`.
+   * - A template string array if the log record was created with a tagged
+   *   template literal syntax, e.g., `["Hello, ", "!"]` for
+   *   ``logger.info`Hello, ${name}!```.
+   *
+   * @since 0.6.0
+   */
+  readonly rawMessage: string | TemplateStringsArray;
+
+  /**
    * The timestamp of the log record in milliseconds since the Unix epoch.
    */
   readonly timestamp: number;